Consolidate tests together and remove redundant tests between page and fileop
