+++ lib/notecli.rb
@@ -1,5 +1,6 @@
 
 require 'notecli/version'
+require 'notecli/fileop'
 require 'notecli/history'
 require 'notecli/cli'
 require 'notecli/config'
