+++ lib/notecli/page.rb
@@ -11,88 +11,17 @@ module Note
   #
   # this class operates to interaction with the file structure, does not store
   # any state data.
-  class Page
-    attr_reader :name, :path
-    
-    def initialize(name, config: Config.new)
-			@config = config.settings
-      self.class.assert
-      self.create name
-    end
-
-    # returns the correct path to the page storage dir
-    def self.path_to(pagename="", config: Config.new)
-      name = self.name.split(':').last.downcase
-      path = "#{name}_path"
-      File.expand_path File.join(config.settings[path], pagename)
-    end
-
-    # asserts that the page storage path exists
-    def self.assert
-      FileUtils.mkdir_p self::path_to
-    end
-
-    def self.exists?(pagename)
-      File.exists? self::path_to(pagename) 
-    end
-
-    # returns page objects for each matching page name
-    def self.find(match="*")
-      self::find_path(match).map{|f| self.new File.basename(f)}
-    end
-
-    # returns full paths for each matching page name
-    def self.find_path(match="*")
-      Dir[(self::path_to match)]
-    end
-
-    def self.search(match="*")
-      found = []
-      self::find.each do |page|
-        # supposedly will be more memory efficient
-        open(page.path) do |f|
-
-          # from The Tin man
-          # https://stackoverflow.com/questions/5761348/ruby-grep-with-line-number
-          grep = f.each_line.with_index(1).inject([]) { |m,i| m << i if (i[0][match]); m }
-
-          if grep.length > 0
-            grep.each do |g|
-              found << {page: page, grep: g.first, line: g.last}
-            end
-          end
-        end
-      end
-      return found || []
-    end
-
-    # creates a symlink to another path for the original path of
-    # this page
-    def symlink(to_path)
-      FileUtils.rm to_path if File.exist? to_path 
-      FileUtils.ln_s(@path, to_path)
-    end
-
-    # runs all of the creation steps for the page dir
-    def create(name)
-      @name = name
-      @path = self.class.path_to name
-      self.touch
-    end
-
-    def touch
-      FileUtils.touch(@path)
-    end
+  class Page < FileOp
 
     # opens a file with the editor and file type provided
     # a file is symlinked to a tmp directory and opened with 
     # a different file extension
     def open(editor: nil, ext: nil)
 
-      editor ||= @config["editor"]
-      ext ||= @config["ext"]
+      editor ||= @settings["editor"]
+      ext ||= @settings["ext"]
 
-      self.class.assert
+      self.class.assert_home
       temp = self.temp ext
       
       system(editor.to_s, temp.to_s)
@@ -111,13 +40,13 @@ module Note
       end
     end
 
-    def self.open_multiple(pages, editor: nil, ext: nil)
+    def self.open_multiple(pages, editor: nil, ext: nil, config: Config.new)
       
-      editor ||= @config["editor"]
-      ext ||= @config["ext"]
+      editor ||= config.settings["editor"]
+      ext ||= config.settings["ext"]
       history = History.new
 
-      self::assert
+      self::assert_home
       temps = pages.map{|page| page.temp ext}
       if temps.length > 0
         conns = "#{editor} #{temps.join(' ')}"
@@ -134,8 +63,8 @@ module Note
     # creates a symlink in a temp directory
     def temp(ext=nil, temp_path=nil)
 
-      temp_path ||= @config["temp_path"]
-      ext ||= @config["ext"]
+      temp_path ||= @settings["temp_path"]
+      ext ||= @settings["ext"]
      
       to_path = File.join(
         temp_path, 
@@ -147,56 +76,11 @@ module Note
     end
 
     # removes a temp file after it is created
-    def rm_temp(file_ext, temp_path: @config["temp_path"])
+    def rm_temp(file_ext, temp_path: @settings["temp_path"])
       path = File.join(temp_path, [@name, file_ext].join("."))
       if File.file? path
         FileUtils.rm path
       end
-    end
-
-    # renames a file to a new name and checks if a file
-    # exists first 
-    def rename(name)
-      FileUtils.mv @path, self.class.path_to(name)
-      self.create name
-      return true
-    end
-
-    # delete the page
-    def delete
-      FileUtils.rm @path
-    end
-
-    # write file
-    def write(data)
-      data = data.join("\n") if data.kind_of?(Array)
-
-      file = File.open(self.path, "w")
-      file.write(data)
-    end
-
-    # appends the file content with a string
-    def append(string)
-      File.open(@path, "a") do |file|
-        file.write string
-      end
-    end
-
-    # prepends the top of a note with a new line
-    def prepend(string)
-			File.open(@path, "r") do |orig|
-					File.unlink(@path)
-					File.open(@path, "w") do |new|
-							new.write string
-							new.write orig.read()
-					end
-			end
-    end
-
-    # returns file contents as string
-    def read
-      file = File.open(@path)
-      file.read
     end
   end
 end
